pyproject.toml
[project]
name = "midas_client"  
version = "1.0.0" 
description = "A Python client library to access the midas_server backend."  
readme = "README.md"  
# readme-content-type = "text/markdown"  
keywords = ["midas", "client", "backend", "API"]  
license = { text = "Apache-2.0" }
# license-files = ["LICENSE"] 
authors = [
    { name = "Anthony Baxter", email = "anthonybaxter819@gmail.com" }
]
requires-python = ">=3.10"  
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent"
]

dependencies = [
    "certifi>=2024.7.4",
    "charset-normalizer>=3.3.2",
    "idna>=3.7",
    "numpy>=2.0.1",
    "packaging>=24.1",
    "pandas>=2.2.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pytz>=2024.1",
    "requests>=2.32.3",
    "six>=1.16.0",
    "tzdata>=2024.1",
    "urllib3>=2.2.2"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

